--- be/src/main/java/codesquad/bookkbookk/domain/Temp.java
@@ -1,13 +1,0 @@
-package codesquad.bookkbookk.domain;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public class Temp {
-
-    private final Long bookClubId;
-    private final String invitationUrl;
-
-}
